hotfix: Refactored to follow SOLID - Implemented Unit tests (Close #16)
Key changes:
- Created utility class.
- Following SOLID.
- Added javadoc.
- Added unit tests.
