+++ src/test/java/com/seyed/ali/timeentryservice/service/TimeEntryServiceImplTest.java
@@ -1,7 +1,7 @@
 package com.seyed.ali.timeentryservice.service;
 
-import com.seyed.ali.sharelib.exceptions.ResourceNotFoundException;
 import com.seyed.ali.timeentryservice.client.AuthenticationServiceClient;
+import com.seyed.ali.timeentryservice.exceptions.ResourceNotFoundException;
 import com.seyed.ali.timeentryservice.model.domain.TimeEntry;
 import com.seyed.ali.timeentryservice.model.domain.TimeSegment;
 import com.seyed.ali.timeentryservice.model.dto.TimeEntryDTO;
@@ -30,6 +30,7 @@ import java.util.UUID;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.catchThrowable;
+import static org.mockito.ArgumentMatchers.isA;
 import static org.mockito.Mockito.*;
 
 @SuppressWarnings("unused")
